Move MoonSpaces to common.
They're not used in common, but this makes Behavior all based on just
Common.
